// watcher/src/core/interfaces.ts
export interface GasTxSettings {
  gasLimit: bigint;
  maxFeePerGas: bigint;
  maxPriorityFeePerGas: bigint;
}
// ...
export interface GasAnalysis {
  gasEstimate: bigint;
  gasCostWEI: bigint;
  gasCostUSD: number;
  nativeTokenPriceUSD: number;
  baseFeePerGas: bigint; // latest base fee from latest block
  gasTxSettings: GasTxSettings;
}

export interface SwapStep {
  pool: DexPoolState;
  tokenIn: TokenOnChain;
  tokenOut: TokenOnChain;
  amountIn: bigint;
  amountOut: bigint;
  spotPrice: number;
  executionPrice: number;
  priceImpact: number;
  slippage: number;
}

// ================================================================================================
// ARBITRAGE OPPORTUNITIES
// ================================================================================================
export interface ArbitrageOpportunity {
  id: string;
  chainId: number;
  status: 'new' | 'invalid' | 'pending' | 'success' | 'failed' | 'error' | 'dropped' | 'timeout';

  // info
  grossProfitToken: bigint; // Profit in borrow token
  grossProfitUSD: number;
  netProfitUSD: number;
  borrowToken: TokenOnChain; // First token in = last token out
  borrowAmount: bigint;

  // details
  steps: SwapStep[];
  gasAnalysis: GasAnalysis;
  logs: any[];

  // Contract Trade Struct
  trade?: Trade; // filled when preparing for execution

  // bribe info
  bribe?: { internalBribeBps: bigint; bribeWEI: bigint; result: any }; // filled when preparing for execution

  // Metrics
  totalSlippage: number;
  totalPriceImpact: number;
  minConfidence?: number;

  // Metadata
  timestamp: number;
  foundAtBlock?: number;
}